# backendApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Punch System API",
        "description": "打卡系統 API 文件",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/departments": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "取得所有部門",
                "description": "此 API 取得所有部門。",
                "operationId": "da8820891dd6bf08713e9d77fbe6389f",
                "responses": {
                    "200": {
                        "description": "成功獲取所有部門",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "成功獲取所有部門"
                                        },
                                        "departments": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Department"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未授權，請提供有效的 Bearer Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "未授權的請求"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Departments"
                ],
                "summary": "新增部門",
                "description": "此 API 新增部門",
                "operationId": "c0f48115550ccf6a76b43ff263bfc05c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "人資部"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "部門已新增",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "部門已新增"
                                        },
                                        "department": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "請求驗證錯誤"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/departments/{id}": {
            "delete": {
                "tags": [
                    "Departments"
                ],
                "summary": "刪除部門",
                "operationId": "5cc7aebc708fbbbeb04b81504f62b499",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "部門 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "部門刪除成功"
                    },
                    "404": {
                        "description": "找不到部門"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Departments"
                ],
                "summary": "更新部門名稱",
                "operationId": "86358cf8bb4f728cb89a0b4225133b63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "部門 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "研發部"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "部門更新成功"
                    },
                    "404": {
                        "description": "找不到部門"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Department": {
                "title": "Department",
                "description": "部門資料結構",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "人事部"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}